Fixes #13 and updates #18
For #13, equipping equipment will append "skip" onto the player's
status. This is then checked in the player's attacking state and handled
accordingly.

For #18, "dead" is now appended to a characters status and removed when
health increases above 0.

General: This new way of handling status will be very powerful and
helpful when we introduce buffs/debuffs. No longer will a character have
simply one status state at a time.